Renamed `item_types.py` to singular `item_type.py` and edited `models/__init__.py` to reflect these changes
