+++ src/api/models/__init__.py
@@ -1,4 +1,4 @@
 from .player import PlayerIn, Player, PlayerTable
 from .item import ItemIn, Item, ItemTable
 from .rarity import Rarity
-from .item_types import ItemTypes
+from .item_type import ItemType
